make docs components more content-agnoistic
