+++ packages/site-kit/components/docs/index.js
@@ -1,0 +1,2 @@
+export { default as Main } from './Main.svelte';
+export { default as Contents } from './Contents.svelte';
