+++ packages/site-kit/index.js
@@ -1,5 +1,4 @@
 export { default as Blurb } from './components/Blurb.svelte';
-export { default as Docs } from './components/Docs.svelte';
 export { default as Hero } from './components/Hero.svelte';
 export { default as Icon } from './components/Icon.svelte';
 export { default as Icons } from './components/Icons.svelte';
