+++ packages/site-kit/package.json
@@ -30,6 +30,7 @@
 		"./base.css": "./base.css",
 		"./code.css": "./code.css",
 		"./components/*": "./components/*",
+		"./components/docs": "./components/docs/index.js",
 		"./fonts/*": "./fonts/*",
 		"./icons/*": "./icons/*",
 		"./utils/*": "./utils/*"
